+++ core-ng/src/main/java/core/framework/module/APIConfig.java
@@ -50,8 +50,8 @@ public class APIConfig extends Config {
                 .trustAll()
                 .connectTimeout(Duration.ofSeconds(2))
                 .timeout(Duration.ofSeconds(20))    // refer to: kube graceful shutdown period is 30s, db timeout is 15s
-                .maxRetries(5)
-                .slowOperationThreshold(Duration.ofSeconds(10));
+                .slowOperationThreshold(Duration.ofSeconds(10))
+                .maxRetries(5);
     }
 
     public <T> void service(Class<T> serviceInterface, T service) {