+++ core-ng/src/main/java/core/framework/module/KafkaConfig.java
@@ -70,7 +70,7 @@ public class KafkaConfig extends Config {
     MessageProducer createProducer() {
         var producer = new MessageProducerImpl(uri, name);
         context.stat.metrics.add(producer.producerMetrics);
-        context.shutdownHook.add(ShutdownHook.STAGE_3, producer::close);
+        context.shutdownHook.add(ShutdownHook.STAGE_4, producer::close);
         var controller = new KafkaController(producer);
         context.route(HTTPMethod.POST, managementPathPattern("/topic/:topic/message/:key"), controller::publish, true);
         return producer;